+++ tests/test_datastructures.py
@@ -815,11 +815,11 @@ class TestCallbackDict:
 
 class TestCacheControl:
     def test_repr(self):
-        cc = ds.RequestCacheControl([("max-age", "0"), ("private", "True")])
+        cc = ds.RequestCacheControl({"max-age": "0", "private": "True"})
         assert repr(cc) == "<RequestCacheControl max-age='0' private='True'>"
 
     def test_set_none(self):
-        cc = ds.ResponseCacheControl([("max-age", "0")])
+        cc = ds.ResponseCacheControl({"max-age": "0"})
         assert cc.no_cache is None
         cc.no_cache = None
         assert cc.no_cache is None
@@ -827,33 +827,33 @@ class TestCacheControl:
         assert cc.no_cache is None
 
     def test_no_transform(self):
-        cc = ds.RequestCacheControl([("no-transform", None)])
+        cc = ds.RequestCacheControl({"no-transform": None})
         assert cc.no_transform is True
         cc = ds.RequestCacheControl()
         assert cc.no_transform is False
 
     def test_min_fresh(self):
-        cc = ds.RequestCacheControl([("min-fresh", "0")])
+        cc = ds.RequestCacheControl({"min-fresh": "0"})
         assert cc.min_fresh == 0
-        cc = ds.RequestCacheControl([("min-fresh", None)])
+        cc = ds.RequestCacheControl({"min-fresh": None})
         assert cc.min_fresh is None
         cc = ds.RequestCacheControl()
         assert cc.min_fresh is None
 
     def test_must_understand(self):
-        cc = ds.ResponseCacheControl([("must-understand", None)])
+        cc = ds.ResponseCacheControl({"must-understand": None})
         assert cc.must_understand is True
         cc = ds.ResponseCacheControl()
         assert cc.must_understand is False
 
     def test_stale_while_revalidate(self):
-        cc = ds.ResponseCacheControl([("stale-while-revalidate", "1")])
+        cc = ds.ResponseCacheControl({"stale-while-revalidate": "1"})
         assert cc.stale_while_revalidate == 1
         cc = ds.ResponseCacheControl()
         assert cc.stale_while_revalidate is None
 
     def test_stale_if_error(self):
-        cc = ds.ResponseCacheControl([("stale-if-error", "1")])
+        cc = ds.ResponseCacheControl({"stale-if-error": "1"})
         assert cc.stale_if_error == 1
         cc = ds.ResponseCacheControl()
         assert cc.stale_while_revalidate is None